+++ utils/Utils.py
@@ -2,11 +2,28 @@ import pickle as pickle
 import os
 import pandas as pd
 import numpy as np
+import torch
+import torch.nn as nn
 import torch.nn.functional as F
 from datetime import datetime
 from datetime import timezone
 from datetime import timedelta
 from glob import glob
+
+class FocalLoss(nn.Module):
+    def __init__(self, weight=None, gamma=0.5, reduction="mean"):
+        nn.Module.__init__(self)
+        self.weight = weight
+        self.gamma = gamma
+        self.reduction = reduction
+
+    def forward(self, input_tensor, target_tensor):
+        log_prob = F.log_softmax(input_tensor, dim=-1)
+        prob = torch.exp(log_prob)
+        return F.nll_loss(
+            ((1 - prob) ** self.gamma) * log_prob, target_tensor, weight=self.weight, reduction=self.reduction
+        )
+
 
 
 def label_to_num(label):
